Move and update get_instruction_encoding tests
