+++ src/vm/run_context.rs
@@ -61,76 +61,6 @@ mod tests {
     use crate::vm::vm_core::VirtualMachineError;
 
     #[test]
-    fn get_instruction_encoding_successful_without_imm() {
-        let mut run_context = RunContext {
-            memory: Memory::new(),
-            pc: MaybeRelocatable::Int(BigInt::from_i32(4).unwrap()),
-            ap: MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-            fp: MaybeRelocatable::Int(BigInt::from_i32(6).unwrap()),
-            prime: BigInt::from_i32(39).unwrap(),
-        };
-
-        run_context.memory.insert(
-            &MaybeRelocatable::Int(BigInt::from_i32(4).unwrap()),
-            &MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-        );
-        if let Ok((num_ref, None)) = run_context.get_instruction_encoding() {
-            assert_eq!(num_ref.clone(), BigInt::from_i32(5).unwrap());
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
-    fn get_instruction_encoding_successful_with_imm() {
-        let mut run_context = RunContext {
-            memory: Memory::new(),
-            pc: MaybeRelocatable::Int(BigInt::from_i32(4).unwrap()),
-            ap: MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-            fp: MaybeRelocatable::Int(BigInt::from_i32(6).unwrap()),
-            prime: BigInt::from_i32(39).unwrap(),
-        };
-
-        run_context.memory.insert(
-            &MaybeRelocatable::Int(BigInt::from_i32(4).unwrap()),
-            &MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-        );
-        run_context.memory.insert(
-            &MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-            &MaybeRelocatable::Int(BigInt::from_i32(6).unwrap()),
-        );
-        if let Ok((num_ref, Some(&MaybeRelocatable::Int(ref imm_addr_num_ref)))) =
-            run_context.get_instruction_encoding()
-        {
-            assert_eq!(num_ref.clone(), BigInt::from_i32(5).unwrap());
-            assert_eq!(imm_addr_num_ref.clone(), BigInt::from_i32(6).unwrap());
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
-    fn get_instruction_encoding_unsuccesful() {
-        let mut run_context = RunContext {
-            memory: Memory::new(),
-            pc: MaybeRelocatable::Int(BigInt::from_i32(4).unwrap()),
-            ap: MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-            fp: MaybeRelocatable::Int(BigInt::from_i32(6).unwrap()),
-            prime: BigInt::from_i32(39).unwrap(),
-        };
-
-        run_context.memory.insert(
-            &MaybeRelocatable::Int(BigInt::from_i32(7).unwrap()),
-            &MaybeRelocatable::Int(BigInt::from_i32(5).unwrap()),
-        );
-        if let Err(error) = run_context.get_instruction_encoding() {
-            assert_eq!(error, VirtualMachineError::InvalidInstructionEncoding);
-        } else {
-            assert!(false);
-        }
-    }
-
-    #[test]
     fn compute_dst_addr_for_ap_register() {
         let instruction = Instruction {
             off0: BigInt::from_i32(1).unwrap(),
